Create main.c
*The project is about the railway ticket booking system. This will include three
destinations: Dhaka,Chittagong and Panchagarh. It takes passengers name and
age . It will use
Different functions for booking, payment, and printing of the ticket. It accepts.
Payment with mobile banking in bKah,Rocket and Nagad and cash. It will use the filling to save the price of the tickets and to.
save the passenger information. It prints the ticket. . The programming.
language of the system is C. we have created the program on Visual Studio Code
and the operating system is Linux.